fix: tab rounded variant
